requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==23.3
django-cors-headers==4.3.1
django-debug-toolbar==4.1.0
psycopg2-binary==2.9.7
boto3==1.28.57
django-storages==1.13.2
celery[redis]==5.3.4
redis==4.6.0
python-dotenv==1.0.0
requests==2.31.0
openai==0.28.0
numpy==1.25.2
pillow==10.0.1
pytesseract
PyPDF2
gunicorn
coreapi==2.3.3
pyyaml==6.0.1
uritemplate==4.1.1